==> 01_dplyr.Rmd <==
---
title: "Dplyr"
author: "Jairo Nicolau"
date: "8/18/2020"
output: html_document
---

### Carregando o banco

DPLYR

```{r}
library(tidyverse) 
library(gapminder)
library(readr)
library(readxl)
library(here)
```

```{r}
View(gapminder)
```

### Operador pipe

CONTROL + SHIFT + M

```{r}
gapminder %>% glimpse 
```

### *Filter* para filtrar linhas

```{r}
gapminder %>% 
  filter (lifeExp > 35) 
```

```{r}
gapminder %>% 
  filter (country == "Argentina", year >= 1992)
```

```{r}
gapminder %>% 
  filter (country %in% c("Rwanda", "Afghanistan", "Brazil"))
```

### *Select* para selecionar colunas

```{r}
gapminder %>%
  select(year, lifeExp, continent) 
```

```{r}
gapminder %>%
  filter(country == "Brazil") %>%
  select(year, lifeExp) %>% 
  top_n(10)
```

### *Arrange* para ordenar por uma variável

```{r}
gapminder %>%
  arrange(year, lifeExp)
```

```{r}
gapminder  %>%
  filter(year == 2007) %>%
  arrange(desc(lifeExp)) %>% 
  select(country, lifeExp) 

```

### *Rename* para renomear as variáveis

```{r}
gapminder %>%
  rename(expectiva_vida = lifeExp,
         gdp_percapita = gdpPercap,
         continente = continent,
         ano = year)
```

### *Mutate* para criar uma nova variável

```{r}
gapminder %>%
  mutate(pib = pop * gdpPercap)  %>% 
  select(pib)
```

#### Contando as categorias

```{r}
gapminder %>%
  count(continent)
  
```

```{r}
gapminder %>%
  count(continent) %>% 
  mutate(prop = n/(sum(n))*100)
  
```

### *Groupby* + *Summarize*

Agregar por uma variavel categorica

```{r}
gapminder %>%
  group_by(continent) %>%
  summarize(casos = n())
```

```{r}
gapminder %>%
  group_by(continent) %>%
  count() %>% 
  ungroup() %>% 
  mutate(prop = n/(sum(n))*100)
```

```{r}
gapminder %>%
  group_by(continent) %>%
  summarize(n = n(),
            n_observacoes = n_distinct(year))
```

#### Sumarização geral

mean(), median(), var(), sd(), IQR(), min(), max()

```{r}
gapminder %>%
  filter(year==2007) %>% 
 group_by(continent) %>%
  summarize(mediana = median(lifeExp),
            media = mean(lifeExp))

```

```{r}
gapminder %>%
  filter(continent == "Asia") %>%
  group_by(year) %>%
  summarize(min_lifeExp = min(lifeExp),
            max_lifeExp = max(lifeExp))
```


#### Exercicio



```{r}
library(readxl) # para ler arquivos excel
library(here) # para ler dados nas pastas

df <- read_excel(here("bairro_rio_2008.xlsx"))
View(df)
```

```{r}
glimpse(df)
```

#### selecionando as variaveis importantes

```{r}
bairro <- df %>% 
  select(Bairro_IPP, GPP, RegAdmin, Dom, DomFav, Pop2000, PopF, RDom_Media, Rlg_EMis_pc,Rlg_EPtc_pc)
```


#### criando variáveis


```{r}
bairro %>% 
mutate(pct_favela = DomFav/ Dom,
       pct_feminino = PopF/Pop2000,
       pct_evangelico = Rlg_EMis_pc + Rlg_EPtc_pc ) -> bairro 
```


#### salvando o banco

```{r}
library(readr)
write_csv(bairro, "bairros_rio.csv")
```
