==> 02_media_desvio.Rmd <==
---
title: "Meidas de tendiencia central e dispersão"
author: "Jairo Nicolau"
date: "01/09/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(here)
```


```{r}
df2 <- read_csv(here("bairros_rio.csv"))
```



### Fazendo sinteses

```{r}
bairro_compacto %>% 
  group_by(GPP) %>%
  summarize(media_favela = mean (pct_favela),
            media_feminino = mean(pct_feminino),
            media_evangelico = mean(pct_evangelico)) 
```




Média e mediana

```{r}
gapminder %>%
  filter(year==2007) %>% 
  group_by(continent) %>%
  summarize(mediana_lifeExp = median (lifeExp), 
            media_lifeExp = mean (lifeExp)) %>% 
  arrange(desc(mediana_lifeExp)) 

```





Sumário de 5 Números




```{r}
df2 %>% 
  summarize( n = n(), 
             Min = min(pct_favela),
             Q1 = quantile(pct_favela, .25),
             Avg = mean(pct_favela), 
             Q3 = quantile(pct_favela, .75),
             Max = max(pct_favela)
             )
```

```{r}
df2 %>% 
   
  summary
```



```{r}
df2 %>% 
select(pct_favela) %>%
  summary()
```

### Encontrando quartis específicos

```{r}
df2 %>% 
  summarize( 
             bottom_10 = quantile(pct_evangelico, .10),
             mediana = mean(pct_favela), 
             top_25 = quantile(pct_evangelico, .75),
             top_1 = quantile(pct_evangelico, .99)
             )
```
